fix size_t conversion warnings in lib/common/memory
    The general strategy here is to promote up to size_t. If it did not make sense to support integers that large for 64 bit
    architectures I added an Assert.

    CustomHeap.cpp
     removed size_t warning
     GetChunkSizeForBytes returns a uint so there is no need for the type to get promoted for x64 arch
     FillDebugBreak now takes in size_t

    CustomHeap.h
     MaxAllocation is a const positive integer so just make it a uint
     promote DecommitPages and AllocSecondary to size_t
     function calls within DecommitPages take size_t already so its safe to promote
     function calls with in AllocSecondary expect a DWORD, add an assert to confirm functionSize_t is not larger
     than MAXUINT32 then explicitly cast down

    HeapBlock.cpp
     remove pragma warning

    heapinfo.cpp
     rescan returns a size_t promote HeapInfo's Rescan to also return a size_t

    heapinfo.h
     change function header to support size_t return type

    LargeHeapBlock.h
     function header update

    LargeHeapBlock.cpp
     remove pragma warning
     LargeHeapBlockFreeListEntry->objectSize returns size_t, all functions that use it take in size_t so I promoted it

    LargeHeapBucket.h
     function header updates

    LargeHeapBucket.cpp
     remove pragma warning
     pass on Rescan pagecount changes on up through return type
     Check() now returns size_t

    PageAllocator.h
     function header updates

    PageAllocator.cpp
     promoted AllocPages & AllocPagesPageAligned, ReleasePages, & DecommitPages to take in pageCounts of size_t
     but also added Asserts b\c it does not make sense to have 4kb * 2^32-1  = 17.18 TB of pages

    Recycler.h
     function header updates

    Recycler.cpp
     passing on heapBlockMap.Rescan size_t through FinishMarkRescan return type

    RecyclerPageAllocator.cpp
     remove pragma warning
     add assert as we don't expect to 17.18 TB of pages

    VirtualAllocWrapper.cpp
     remove pragma warning
     requestedNumOfSegments is not expected to be larger than MAXUINT32 so Assert and static cast down

+++ lib/common/Memory/LargeHeapBlock.h
@@ -210,7 +210,7 @@ private:
     template <SweepMode>
     void SweepObject(Recycler * recycler, LargeObjectHeader * header);
 
-    bool TrimObject(Recycler* recycler, LargeObjectHeader* header, uint sizeOfObject, bool needSuspend = false);
+    bool TrimObject(Recycler* recycler, LargeObjectHeader* header, size_t sizeOfObject, bool needSuspend = false);
 
     void FinalizeObject(Recycler* recycler, LargeObjectHeader* header);
